requirements.txt
asgiref
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
curtsies==0.3.0
dj-database-url==0.5.0
Django==3.1.2
django-filter==2.3.0
django-nose==1.4.6
django-rest-passwordreset==1.1.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.10.0
djangorestframework==3.12.4
greenlet==0.4.15
gunicorn==19.9.0
itypes==1.1.0
MarkupSafe==1.1.0
nose==1.3.7
openapi-codec==1.3.2
Pillow==5.4.1
PyJWT==1.7.1
requests
simplejson==3.16.0
typing==3.6.6
uritemplate==3.0.0
urllib3==1.22
whitenoise==5.3.0
google-api-python-client==1.12.8
jet-django
django-phonenumber-field==2.2.0
phonenumbers==8.10.4
pandas
rest-condition
djangorestframework-csv
django-storages
celery==5.0.2
redis==3.2.0
djangorestframework-simplejwt
coverage-badge
pyUFbr
psycopg2-binary
openpyxl
pycep-correios
django-environ==0.4.5
googlemaps
daphne==3.0.0
